User Policy

+++ api/app/Http/Controllers/Api/V1/UserController.php
@@ -12,6 +12,7 @@ use App\Http\Requests\User\UpdateRequest;
 use App\Http\Resources\User\IndexResource;
 use App\Http\Resources\User\ShowResource;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -30,6 +31,8 @@ class UserController extends Controller
      */
     public function store(StoreRequest $request, StoreAction $action)
     {
+        Gate::authorize('create', User::class);
+
         $movie = $action($request);
 
         return new ShowResource($movie);
@@ -48,6 +51,8 @@ class UserController extends Controller
      */
     public function update(UpdateRequest $request, User $user, UpdateAction $action)
     {
+        Gate::authorize('update', $user);
+
         $updatedUser = $action($request, $user);
 
         return new ShowResource($updatedUser);
@@ -58,6 +63,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        Gate::authorize('delete', $user);
+
         $user->delete();
 
         return response()->noContent();
@@ -65,6 +72,8 @@ class UserController extends Controller
 
     public function updateAvatar(AvatarUpdateRequest $request, User $user, AvatarUpdateAction $action) 
     {
+        Gate::authorize('update', $user);
+
         $updatedUser = $action($request, $user);
 
         return new ShowResource($updatedUser);
